+++ src/coreaudio_backend/device.rs
@@ -14,6 +14,7 @@
 // Mozilla Public License along with this program. If not, see
 // <https://www.gnu.org/licenses/> and <https://www.mozilla.org/MPL/2.0/>.
 
+use crate::ToF32;
 use crate::audiochunk::ChunkStats;
 use crate::audiodevice::*;
 use crate::config;
@@ -1156,7 +1157,7 @@ impl CaptureDevice for CoreaudioCaptureDevice {
                             averager.restart();
                             if let Ok(mut capture_status) = RwLockUpgradableReadGuard::try_upgrade(capture_status) {
                                 capture_status.measured_samplerate = measured_rate as usize;
-                                capture_status.signal_range = value_range as f32;
+                                capture_status.signal_range = value_range.to_f32();
                                 capture_status.rate_adjust = rate_adjust as f32;
                                 crate::update_capture_state(&mut capture_status, state);
                             }