+++ src/file_backend/device.rs
@@ -14,8 +14,7 @@
 // Mozilla Public License along with this program. If not, see
 // <https://www.gnu.org/licenses/> and <https://www.mozilla.org/MPL/2.0/>.
 
-#![cfg_attr(camillafloat_f32, allow(clippy::unnecessary_cast))]
-
+use crate::ToF32;
 use crate::audiochunk::{AudioChunk, ChunkStats};
 use crate::audiodevice::*;
 use crate::config;
@@ -559,7 +558,7 @@ fn capture_loop(
                             RwLockUpgradableReadGuard::try_upgrade(capture_status)
                         {
                             capture_status.measured_samplerate = measured_rate as usize;
-                            capture_status.signal_range = value_range as f32;
+                            capture_status.signal_range = value_range.to_f32();
                             capture_status.rate_adjust = rate_adjust as f32;
                             crate::update_capture_state(&mut capture_status, state);
                         } else {